network: Move unprivileged setup to its own package
virt-handler (privileged) and virt-launcher (unprivileged) network
setup shared pkg/network/setup/. Move the virt-launcher part into a
child pkg/network/setup/launcher package to make the boundary
explicit.
