+++ pkg/network/setup/launcher/BUILD.bazel
@@ -3,58 +3,47 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "go_default_library",
     srcs = [
-        "configstatecache.go",
-        "filters.go",
-        "netconf.go",
-        "netstat.go",
+        "network.go",
+        "podnic.go",
     ],
-    importpath = "kubevirt.io/kubevirt/pkg/network/setup",
+    importpath = "kubevirt.io/kubevirt/pkg/network/setup/launcher",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/network/cache:go_default_library",
-        "//pkg/network/deviceinfo:go_default_library",
+        "//pkg/network/dhcp:go_default_library",
+        "//pkg/network/domainspec:go_default_library",
         "//pkg/network/driver:go_default_library",
-        "//pkg/network/istio:go_default_library",
-        "//pkg/network/netns:go_default_library",
-        "//pkg/network/setup/netpod:go_default_library",
-        "//pkg/network/setup/netpod/masquerade:go_default_library",
+        "//pkg/network/link:go_default_library",
         "//pkg/network/vmispec:go_default_library",
-        "//pkg/util:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
-        "//pkg/virt-launcher/virtwrap/converter/network:go_default_library",
-        "//pkg/vmitrait:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/types:go_default_library",
-        "//vendor/k8s.io/utils/net:go_default_library",
+        "//staging/src/kubevirt.io/client-go/precond:go_default_library",
     ],
 )
 
 go_test(
     name = "go_default_test",
     srcs = [
-        "configstatecache_test.go",
-        "filters_test.go",
-        "netconf_test.go",
-        "netstat_test.go",
-        "network_suite_test.go",
+        "launcher_suite_test.go",
+        "network_test.go",
+        "podnic_test.go",
     ],
     embed = [":go_default_library"],
     race = "on",
     deps = [
         "//pkg/ephemeral-disk-utils:go_default_library",
         "//pkg/libvmi:go_default_library",
-        "//pkg/libvmi/status:go_default_library",
         "//pkg/network/cache:go_default_library",
-        "//pkg/network/deviceinfo:go_default_library",
-        "//pkg/network/setup/netpod:go_default_library",
-        "//pkg/network/vmispec:go_default_library",
+        "//pkg/network/dhcp:go_default_library",
+        "//pkg/network/driver:go_default_library",
+        "//pkg/network/namescheme:go_default_library",
         "//pkg/os/fs:go_default_library",
         "//pkg/virt-launcher/virtwrap/api:go_default_library",
         "//staging/src/kubevirt.io/api/core/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/testutils:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
+        "//vendor/go.uber.org/mock/gomock:go_default_library",
     ],
 )
