+++ pkg/network/setup/launcher/podnic.go
@@ -17,7 +17,7 @@
  *
  */
 
-package network
+package launcher
 
 import (
 	v1 "kubevirt.io/api/core/v1"
