+++ pkg/network/setup/launcher/podnic_test.go
@@ -16,7 +16,7 @@
  * Copyright The KubeVirt Authors.
  */
 
-package network
+package launcher
 
 import (
 	"fmt"
