+++ pkg/virt-launcher/virtwrap/BUILD.bazel
@@ -31,7 +31,7 @@ go_library(
         "//pkg/liveupdate/memory:go_default_library",
         "//pkg/network/cache:go_default_library",
         "//pkg/network/deviceinfo:go_default_library",
-        "//pkg/network/setup:go_default_library",
+        "//pkg/network/setup/launcher:go_default_library",
         "//pkg/network/vmispec:go_default_library",
         "//pkg/os/disk:go_default_library",
         "//pkg/pointer:go_default_library",
